requirements.txt
alembic==1.14.1
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.44.1
Deprecated==1.2.18
docker==7.1.0
Flask==3.1.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mlflow==2.20.2
mlflow-skinny==2.20.2
numpy==2.2.3
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
